Strings
In the first line print two space-separated integers, representing the length of  and  respectively.
In the second line print the string produced by concatenating  and  ().
In the third line print two strings separated by a space,  a' and  b'. a' and b' are the same as a and b, respectively, except that their first characters are swapped.